OpenCL on Win32. Pretty easy! - nVidia driver demands that you specify a platform first (presumably so you'll make sure and pick theirs?) - Now lets the api decide which device type is best. With nv, it's GPU-only. (surprise.) Aparently, AMD supports CPUs. - etc. Time for a new download!
git-svn-id: https://crag.googlecode.com/svn/trunk@61 03cb304c-929a-b8d2-5531-aad0e9782889

+++ src/crag/cl/Kernel.cpp
@@ -88,7 +88,7 @@ char * cl::Kernel::LoadClFile(char const * filename)
 	}
 
 	// Get the file size and create a buffer big enough to load it.
-	size_t file_size = in_file.tellg();
+	size_t file_size = static_cast<size_t>(in_file.tellg());
 	char * kernel_source = new char [file_size + 1];
 	
 	// Load the file as a string.
